requirements.txt
Django==4.1.1
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
firebase-admin==5.3.0
